+++ Assets/Scripts/PlayerStats.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PlayerStats
 {
-    public int baseDamage;
-    public int arrowDamage;
-    public int rockDamage;
+    public long baseDamage;
+    public long arrowDamage;
+    public long rockDamage;
 }
